mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "runestone-mcp-server"
version = "1.0.0"
description = "MCP server for parsing and validating RUNE specifications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
runestone-mcp = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.wheel.sources]
"." = "mcp_server"
